requirements.txt
ultralytics
pillow
streamlit
opencv-python